generate map working nicely so far. Fully reworked the placement logic and loving it.
